#63 Updated Profile Creation with JWT Token
