+++ backend-api/config.js
@@ -1,9 +1,11 @@
+require('dotenv').config();
+
 //Server configuration
 const config = {
-    user: 'SacStateLifeAdmin',
-    password: 'TheNest2024',
-    server: 'csuslife.database.windows.net',
-    database: 'csusLifeAppMainDB',
+    user: process.env.DB_USER,
+    password: process.env.DB_PASSWORD,
+    server: process.env.DB_SERVER,
+    database: process.env.DB_DATABASE,
     options: {
         encrypt: true
     }
